try on cpu

+++ src/metrics/distances.py
@@ -227,7 +227,7 @@ def get_far_clp(id_dict: FeatureDict, ood_dict: FeatureDict, clip_model, tempera
 
 def get_mmd_rbf_kernel(id_features, ood_features):
     import math
-    X = torch.cat((id_features, ood_features)).cpu()
+    X = torch.cat((id_features, ood_features)).to(torch.float32).cpu()
     print(X.shape)
     cd = torch.cdist(X, X)
     print(cd)
